+++ .env.example
@@ -1,1 +1,1 @@
-PORT=
+REACT_APP_URL=
