--- src/components/social/socialLogin.scss
@@ -1,11 +1,0 @@
-
-.social-btn{
-  background-color: transparent;
-  border: transparent;
-  cursor: pointer;
-}
-.social-btn img{
-  height: 50px;
-  width: 50px;
-  margin-left: 20px;
-}
