+++ src/redux/reducers/rootReducer.js
@@ -6,7 +6,6 @@ import passwordReducer from './password.reducer';
 
 const rootReducer = combineReducers({
 
-  requests: requestsReducer,
   login: loginReducer,
   reset: passwordReducer,
   requests: requestsReducer
